worker/requirements.txt
amqp==5.0.6
argcomplete==1.10.3
backports.zoneinfo==0.2.1
beautifulsoup4==4.8.2
billiard==3.6.4.0
blis==0.7.4
catalogue==2.0.6
celery==5.1.2
certifi==2021.5.30
chardet==3.0.4
charset-normalizer==2.0.6
click==7.1.2
click-didyoumean==0.0.3
click-plugins==1.1.1
click-repl==0.2.0
compressed-rtf==1.0.6
cymem==2.0.5
docx2txt==0.8
ebcdic==1.1.1
EbookLib==0.17.1
extract-msg==0.28.7
flower
idna==3.2
IMAPClient==2.1.0
Jinja2==3.0.1
kombu==5.1.0
lxml==4.6.3
MarkupSafe==2.0.1
murmurhash==1.0.5
numpy==1.21.2
olefile==0.46
packaging==21.0
pathy==0.6.0
pdfminer.six
pdfplumber
Pillow==8.3.2
pipdeptree==2.1.0
preshed==3.0.5
prompt-toolkit==3.0.20
pycryptodome==3.10.4
pydantic==1.8.2
pyparsing==2.4.7
python-pptx==0.6.21
pytz==2021.1
redis
regex==2021.9.24
requests==2.26.0
sentry-sdk==1.4.1
shortuuid==1.0.1
six==1.12.0
smart-open==5.2.1
sortedcontainers==2.4.0
soupsieve==2.2.1
spacy==3.1.3
spacy-legacy==3.0.8
SpeechRecognition==3.8.1
srsly==2.4.1
textract
thinc==8.0.10
tqdm==4.62.3
typer==0.4.0
typing-extensions==3.10.0.2
tzlocal==3.0
urllib3==1.26.7
vine==5.0.0
Wand==0.6.7
wasabi==0.8.2
wcwidth==0.2.5
xlrd==1.2.0
XlsxWriter==3.0.1
